requirements.txt
numpy
google-cloud
opencv-python